# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: JJK RPG Dashboard API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /characters/{id}:
    get:
      operationId: getCharacter
      tags: [characters]
      summary: Get character by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: number
      responses:
        "200":
          description: Character
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Character"
        "404":
          description: Not found

  # ── SHOP ──────────────────────────────────────────
...
  /inventory/{itemId}/use:
    post:
      operationId: useItem
      tags: [inventory]
      summary: Use a consumable item
      parameters:
        - name: itemId
          in: path
          required: true
          schema:
            type: number
      responses:
        "200":
          description: Item used
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UseItemResult"
        "400":
          description: Cannot use item
  /inventory/{itemId}/equip:
    post:
      operationId: equipItem
      tags: [inventory]
      summary: Equip a weapon or artifact
      parameters:
        - name: itemId
          in: path
          required: true
          schema:
            type: number
      responses:
        "200":
          description: Item equipped
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UseItemResult"
  /inventory/{itemId}/sell:
    delete:
      operationId: sellItem
      tags: [inventory]
      summary: Sell an item for Yen
      parameters:
        - name: itemId
          in: path
          required: true
          schema:
            type: number
      responses:
        "200":
          description: Item sold
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SellItemResult"
        "404":
          description: Item not in inventory

  # ── DAILY ─────────────────────────────────────────
...
  /casino/coinflip:
    post:
      operationId: playCoinFlip
      tags: [casino]
      summary: Play coin flip
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CoinFlipInput"
      responses:
        "200":
          description: Result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CasinoResult"
  /casino/dice:
    post:
      operationId: playDice
      tags: [casino]
      summary: Play dice roll
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CasinoInput"
      responses:
        "200":
          description: Result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CasinoResult"
  /casino/slots:
    post:
      operationId: playSlots
      tags: [casino]
      summary: Play slot machine
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CasinoInput"
      responses:
        "200":
          description: Result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CasinoResult"
  /casino/roulette:
    post:
      operationId: playRoulette
      tags: [casino]
      summary: Play roulette
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RouletteInput"
      responses:
        "200":
          description: Result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CasinoResult"
  /casino/crash:
    post:
      operationId: playCrash
      tags: [casino]
      summary: Play crash game
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CasinoInput"
      responses:
        "200":
          description: Result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CasinoResult"
...
components:
  schemas:
...
    Character:
      type: object
      required: [id, name, grade, technique, attack, defense, speed, max_hp, max_ce, cost]
      properties:
        id:
          type: number
        name:
          type: string
        grade:
          type: string
        quote:
          type: ["string", "null"]
        technique:
          type: string
        attack:
          type: number
        defense:
          type: number
        speed:
          type: number
        max_hp:
          type: number
        max_ce:
          type: number
        cost:
          type: number
        image_url:
          type: ["string", "null"]
        attacks:
          type: array
          items:
            $ref: "#/components/schemas/CharacterAttack"
    CharacterAttack:
      type: object
      required: [num, name, ce_cost, dmg_mult, description]
      properties:
        num:
          type: number
        name:
          type: string
        ce_cost:
          type: number
        dmg_mult:
          type: number
        description:
          type: string
        effect:
          type: ["string", "null"]

    # ── SHOP ──────────────────────────────────────────
...
    UseItemResult:
      type: object
      required: [success, message]
      properties:
        success:
          type: boolean
        message:
          type: string
        hp_gained:
          type: ["integer", "null"]
        ce_gained:
          type: ["integer", "null"]
        xp_gained:
          type: ["integer", "null"]
        new_hp:
          type: ["integer", "null"]
        new_ce:
          type: ["integer", "null"]
        new_xp:
          type: ["integer", "null"]
    SellItemResult:
      type: object
      required: [success, message, yen_gained, new_yen]
      properties:
        success:
          type: boolean
        message:
          type: string
        yen_gained:
          type: number
        new_yen:
          type: number

    # ── DAILY ─────────────────────────────────────────
...
    CasinoInput:
      type: object
      required: [bet]
      properties:
        bet:
          type: number
    CoinFlipInput:
      type: object
      required: [bet, choice]
      properties:
        bet:
          type: number
        choice:
          type: string
    RouletteInput:
      type: object
      required: [bet, bet_type]
      properties:
        bet:
          type: number
        bet_type:
          type: string
        bet_value:
          type: ["string", "null"]
    CasinoResult:
      type: object
      required: [won, payout, new_yen, game, outcome_text]
      properties:
        won:
          type: boolean
        payout:
          type: number
        new_yen:
          type: number
        game:
          type: string
        outcome_text:
          type: string
        details:
          type: object
          additionalProperties: true
...