# lib/api-zod/src/generated/types/characterAttack.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * JJK RPG Dashboard API
 * OpenAPI spec version: 0.1.0
 */

export interface CharacterAttack {
  num: number;
  name: string;
  ce_cost: number;
  dmg_mult: number;
  description: string;
  /** @nullable */
  effect?: string | null;
}